Crate mutex_sleep

Locking types that put a task to sleep while it waits for the lock.

These are Theseus-specific locking types that ensure mutual exclusion using [spin::Mutex] and [spin::RwLock] under the hood; see those types for more details on how they work.

Structs

MutexSleep

A mutual exclusion wrapper that puts a Task to sleep while waiting for the lock to become available.

MutexSleepGuard

A guard that allows the locked data to be accessed, during which mutual exclusion is guaranteed.

RwLockSleep

A multi-reader, single-writer mutual exclusion wrapper that puts a Task to sleep while waiting for the lock to become available.

RwLockSleepReadGuard

A guard that allows the locked data to be immutably accessed, during which mutual exclusion is guaranteed.

RwLockSleepWriteGuard

A guard that allows the locked data to be mutably accessed, during which mutual exclusion is guaranteed.

Type Definitions

MutexSleepGuardRef

Typedef of a owning reference that uses a MutexSleepGuard as the owner.

MutexSleepGuardRefMut

Typedef of a mutable owning reference that uses a MutexSleepGuard as the owner.

RwLockSleepReadGuardRef

Typedef of a owning reference that uses a RwLockSleepSafeReadGuard as the owner.

RwLockSleepWriteGuardRefMut

Typedef of a mutable owning reference that uses a RwLockSleepWriteGuard as the owner.