Struct mutex_sleep::MutexSleep

pub struct MutexSleep<T: ?Sized> { /* private fields */ }

A mutual exclusion wrapper that puts a Task to sleep while waiting for the lock to become available.

A sleeping Task has a “blocked” runstate, meaning that it will not be scheduled in. Once the lock becomes available, Tasks that are sleeping while waiting for the lock will be notified (woken up) so they can attempt to acquire the lock again.

Implementations

impl<T> MutexSleep<T>

pub fn new(data: T) -> MutexSleep<T>

Creates a new lock wrapping the supplied data.

pub fn into_inner(self) -> T

Consumes this MutexSleep, returning the underlying data.

impl<T: ?Sized> MutexSleep<T>

pub fn is_locked(&self) -> bool

Returns true if the lock is currently held.

Safety

This function provides no synchronization guarantees and so its result should be considered ‘out of date’ the instant it is called. Do not use it for synchronization purposes. However, it may be useful as a heuristic.

pub fn lock(&self) -> Result<MutexSleepGuard<'_, T>, &'static str>

Blocks until the lock is acquired by putting this Task to sleep until another Task that has the lock releases it.

The returned guard may be dereferenced to access the protected data; the lock will be released when the returned guard falls out of scope and is dropped.

pub fn try_lock(&self) -> Option<MutexSleepGuard<'_, T>>

Tries to lock the MutexSleep. If it is already locked, it will return None. Otherwise it returns a guard within Some.

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the MutexSleep mutably, and a mutable reference is guaranteed to be exclusive in Rust, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist. As such, this is a ‘zero-cost’ operation.

Example

let mut lock = MutexSleep::new(0);
*lock.get_mut() = 10;
assert_eq!(*lock.lock(), 10);