Struct mutex_sleep::RwLockSleep

pub struct RwLockSleep<T: ?Sized> { /* private fields */ }

A multi-reader, single-writer mutual exclusion wrapper that puts a Task to sleep while waiting for the lock to become available.

The behavior of this read-write lock is defined by the underlying [spin::RwLock];

A sleeping Task has a “blocked” runstate, meaning that it will not be scheduled in. Once the lock becomes available, Tasks that are sleeping while waiting for the lock will be notified (woken up) so they can attempt to acquire the lock again.

Implementations

impl<T> RwLockSleep<T>

pub fn new(data: T) -> RwLockSleep<T>

Creates a new lock wrapping the supplied data.

pub fn into_inner(self) -> T

Consumes this RwLockSleep, returning the underlying data.

impl<T: ?Sized> RwLockSleep<T>

pub fn is_locked(&self) -> bool

Returns true if the lock is currently held.

Safety

This function provides no synchronization guarantees and so its result should be considered ‘out of date’ the instant it is called. Do not use it for synchronization purposes. However, it may be useful as a heuristic.

pub fn read<'a>(&'a self) -> Result<RwLockSleepReadGuard<'a, T>, &'static str>

Locks this RwLockSleep with shared read (immutable) access.

If a writer has already acquired this lock, this function will put the current Task to sleep until the writer Task releases the lock, at which point this Task will be woken up and attempt to acquire the read-only lock again.

There may be other readers currently inside the lock when this function returns. This does not provide any guarantees with respect to the ordering of whether contentious readers or writers will acquire the lock first.

Returns an RAII guard which will release this task’s shared access upon being dropped.

let mylock = RwLockSleep::new(0);
{
    let mut data = mylock.read();
    // The lock is now locked and the data can be read
    println!("{}", *data);
    // The lock is dropped and interrupts are restored to their prior state
}

pub fn try_read(&self) -> Option<RwLockSleepReadGuard<'_, T>>

Attempt to acquire this lock with shared read (immutable) access.

This function is the same as RwLockSleep::read but will never block, returning immediately regardless of whether the lock has been acquired.

let mylock = RwLockSleep::new(0);
{
    match mylock.try_read() {
        Some(data) => {
            // The lock is now locked and the data can be read
            println!("{}", *data);
            // The lock is dropped
        },
        None => (), // failed, another task holds the writer lock
    };
}

pub fn reader_count(&self) -> usize

Return the number of readers that currently hold the lock (including upgradable readers).

Safety

This function provides no synchronization guarantees and so its result should be considered ‘out of date’ the instant it is called. Do not use it for synchronization purposes. However, it may be useful as a heuristic.

pub fn writer_count(&self) -> usize

Return the number of writers that currently hold the lock.

Because RwLockSleep guarantees exclusive mutable access, this function may only return either 0 or 1.

Safety

This function provides no synchronization guarantees and so its result should be considered ‘out of date’ the instant it is called. Do not use it for synchronization purposes. However, it may be useful as a heuristic.

pub unsafe fn force_read_decrement(&self)

Force decrement the reader count.

This is extremely unsafe if there are outstanding RwLockSleepReadGuards live, or if called more times than read has been called, but can be useful in FFI contexts where the caller doesn’t know how to deal with RAII.

pub unsafe fn force_write_unlock(&self)

Force unlock exclusive write access.

This is extremely unsafe if there are outstanding RwLockSleepWriteGuards live, or if called when there are current readers, but can be useful in FFI contexts where the caller doesn’t know how to deal with RAII.

pub fn write<'a>(&'a self) -> Result<RwLockSleepWriteGuard<'a, T>, &'static str>

Locks this RwLockSleep with exclusive write (mutable) access.

If another writer or reader has already acquired this lock, this function will put the current Task to sleep until all other readers and the other writer Task release their locks, at which point this Task will be woken up and attempt to acquire the write lock again.

This does not provide any guarantees with respect to the ordering of whether contentious readers or writers will acquire the lock first.

Returns an RAII guard which will release this task’s exclusive access upon being dropped.

let mylock = RwLockSleep::new(0);
{
    let mut data = mylock.write();
    // The lock is now locked and the data can be written
    *data += 1;
    // The lock is dropped
}

pub fn try_write(&self) -> Option<RwLockSleepWriteGuard<'_, T>>

Attempt to acquire this lock with exclusive write (mutable) access.

This function is the same as RwLockSleep::write but will never block, returning immediately regardless of whether the lock has been acquired.

let mylock = RwLockSleep::new(0);
{
    match mylock.try_write() {
        Some(mut data) => {
            // The lock is now locked and the data can be written
            *data += 1;
            // The lock is implicitly dropped
        },
        None => (), // failed, another task holds the writer lock
    };
}

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the RwLockSleep mutably, and a mutable reference is guaranteed to be exclusive in Rust, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist. As such, this is a ‘zero-cost’ operation.

Example

let mut lock = RwLockSleep::new(0);
*lock.get_mut() = 10;
assert_eq!(*lock.lock(), 10);